Improved documentation, added goDirectToCourse, changed certs.
Removed the old dynactive test certs and changed them to the new idp generated certs.

Improved the readme documentation and documentation across the board.

Validation enhancement for the course ids for the users.

Handled the goDirectToCourse change.

Put back in the backwards compatibile saml parameter that will disappear when we go live.